Rename the WIT file
